requirements.txt
unstructured[all-docs]
unstructured-client 
watermark 
python-dotenv 
pydantic 
langchain 
langchain-community
langchain_core 
langchain_openai 
chromadb
openai
langchain-groq
streamlit

